clippy module inception: move actor/actor and actor/mod into actor
